requirement.txt
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.5.0
django-discover-runner==1.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-role-permissions==3.1.1
django-taggit==1.3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0